Restructure trading tests to match trading contracts directory layout
